requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.30.0
plotly>=5.18.0
